pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptly-ai"
version = "0.1.0"
description = "A Python package for managing prompts using Jinja2 templates"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["prompts", "templates", "jinja2"]
authors = [
    { name = "Owain Lewis", email = "owain@owainlewis.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "Jinja2>=3.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ipykernel>=6.0.0",
    "types-PyYAML>=6.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/owainlewis/promptly"
Documentation = "https://github.com/owainlewis/promptly#readme"
Repository = "https://github.com/owainlewis/promptly.git"

[tool.hatch.build.targets.wheel]
packages = ["promptly_ai"] 